+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * A `HaveVersion` method to the `LibHeifInfo` class.
   * This simplifies checking that LibHeif is at least the specified minimum version. 
 * `IccColorProfile` and `NclxColorProfile` properties to the `HeifImageHandle` class.
+* A `GetMetadataBlockInfo` method to the `HeifImageHandle` class.
 
 ## Removed